Optimize the structure of AbiLoader, can read user-defined Abi
