+++ coinlib/src/main/java/com/keystone/coinlib/abi/ABIStoreEngine.java
@@ -1,0 +1,6 @@
+package com.keystone.coinlib.abi;
+
+public interface ABIStoreEngine {
+    Contract load();
+    ABIStoreEngine getNext();
+}
