+++ coinlib/src/main/java/com/keystone/coinlib/utils/FileUtil.java
@@ -37,22 +37,4 @@ public class FileUtil {
         }
         return stringBuilder.toString();
     }
-
-    public static String externalSDCardPath() {
-        String sdCardPath = "";
-        try {
-            StorageManager storageManager = (StorageManager) Coinlib.sInstance.getContext().getSystemService(Context.STORAGE_SERVICE);
-
-            // Android N started to have this method
-            List<StorageVolume> storageVolumes = storageManager.getStorageVolumes();
-            Class<?> volumeClass = Class.forName("android.os.storage.StorageVolume");
-            Method getPath = volumeClass.getDeclaredMethod("getPath");
-            getPath.setAccessible(true);
-            StorageVolume storageVolume = storageVolumes.get(storageVolumes.size() - 1);
-            sdCardPath = (String) getPath.invoke(storageVolume);
-        } catch (Exception e) {
-            e.printStackTrace();
-        }
-        return sdCardPath;
-    }
 }
